Update Leviathan to c8b947e; clarify that Mutagen is not imported by the Web app; update copyright notices; rename some license/notice files
WebLeviathan prior to this commit should have been licensed under the
GNU GPL version 2 (or GPLv3+ after Mutagen was relicensed) because
Leviathan unconditionally imported Mutagen.  The Leviathan commit prior
to 575de03 adds an `enable_gpl()` function which allows Mutagen to be
imported.  Since WebLeviathan has no need to use Mutagen-dependent
features, it does not call that function when importing Leviathan,
and so Mutagen is not imported.  This allows WebLeviathan's license
to remain the X11 License.

My apologies to the Mutagen developers.

This commit also incorporates these changes to Leviathan:

* `leviathan.yml` is now the preferred name for the config file; the old
  name (`leviathan.yaml`) still works if the former does not exist.

* Ogg files with `.opus` extensions are supported for adding and
  converting to MP3.  (Not applicable to WebLeviathan.)